Make cursor move and be able to put pieces
Another big commit coming. With this now you can move around the board and
put pieces in it (the pieces aren't centered so there is that problem), the
game doesn't print those boards who are already won by someone neither does
it print info like the current turn or names of players.