Refactor way to handle ESCAPE and updateWinner
Hopefully with this changes it is more clear how pressing ESCAPE
closes the game.

Also fixes a bug where you need to press ESCAPE two times to exit
the game and in game vs AI it closed the menu too.

Also I noticed I never updated the winner variable in the class
TicTacToe so games never ended even if they should. it should be
fixed now.